Add QrDataProvider and move functions to helper file.
